Add Windows script for launching the app locally for devs
Because the JavaScript code gets data.json from the local file system when opening index.html locally, you will get a Cross-Origin Request Block due to the Same Origin Policy if you open it normally.

This little BAT script opens the app in Chrome with the --allow-file-access-from-files flag, which temporarily disables the security measure, allowing the app to run on the local system.