Fix C2E20-C2E49, except C2E29, C2E31, C2E37
First fruits of autosync!

The algorithm had trouble with three episodes, which are not included in this commit:

C2E29 & C2E37 are live shows with long segments of white noise (crowd cheering) at the start, which made the first timestamp inaccurate.

The start of C2E31 was identified correctly, but this episode is odd in that it violates the assumption that the duration of the first segment (everything before break) is the same for both media.