Fix C2E29, C2E31, C2E37
I used autosync to help with after-break times, but I had to make some manual adjustments / ignore its suggested changes for the before-break times.

In the cases of C2E29 and C2E37, this was due to them being live shows with lots of white noise (cheering crowd) at the beginning, which caused matching to fail.

For C2E31, I discovered this episode actually has a couple discontinuities, where a few seconds are missing from the YouTube video in two places.