Update Spotify and Google Podcasts URLs after republishing
Podcast episode GUIDs for C3E15-17 changed in PR #48. Consequently, Spotify and Google Podcasts relisted these episodes with new IDs and new URLs.

Note that C3E16 is now double listed in the podcast feed. If one of these entries is deleted from the feed in the future, the URLs for this episode may need to be updated again.