Fix C3E16 podcast URL and C2E55 timestamps
The lastest podcast feed update (PR #49) removed a duplicate entry for C3E16 (introduced in PR #48), so the Spotify and Google Podcasts URLs stored in data.json needed to be updated.

The podcast for C2E55 was previously (PR #48) published with the incorrect audio (from C2E56). This was corrected with the latest podcast feed update (PR #49) by changing the file available at the same media URL. This allowed the timestamps to be restored and updated for C2E55. However, as of this moment, Spotify still has the old, incorrect audio file cached, since the episode's GUID and URL did not change in the feed.