Add C3E42
Issue #93 applies to C3E42 as well.

The last two podcast timestamps needed to be obtained manually because the episode duration is reported inconsistently.

This episode's duration is reported as 3:41:14 by the podcast feed, and by Spotify in both Chrome and Firefox.


The duration is reported as 3:44:29 by VLC and Windows Explorer file properties.


The duration is reported as 3:44:30 by `librosa.get_duration`, and by Google Podcasts in Chrome and Firefox.